Rewrite
The previous draft would not have worked with any higher class StreamReader, because there are no handshakes for pausing and resuming the underlying transports (which results in RuntimeErrors). So I guess a generic throttled reader is not that easy to implement. For now, I will replace the specific reader class that I want throttled, which is the aiohttp.FlowControlStreamReader.